Add support for better API debugging; start v1.5.0 development
In prepration for #305, this adds a mechanism to inspect all API responses
received by the client. That way if you want to figure out why something isn't
working as expected, or you'd like to make use of an undocumented field, you
have access to the response and can use it.

This also adds a `Do()` method, with the same signature as `*http.Client.Do()`,
that allows consumrs to generate their own request, have the client add
authentication details and other headers, before sending the request to the
PagerDuty API. This allows consumers to further debug and make use of features
not yet supported in the REST client.

Also, updates Version string to v1.5.0 as this is work towards that release.

Lastly, this updates some of the test files to remove some deduplication across
them. These changes were needed due to the new fields added to the struct that
had to be initialized to make HTTP requests.